refactored some of the utility functions out of dashboard summary.
