+++ public/javascripts/create-dom-elements.js
@@ -156,8 +156,13 @@ function decorateSearchItem(div, task) {
     });
 };
 
-// create task summary
+
 export async function buildTaskSummary(task) {
+    /*
+    This function gets called every time a task is selected in the task inbox.
+    It builds the task summary panel and will clear the previous contents each time a different task is selected.
+    */
+
     const currentTask = task.name;
     const currentTaskId = task.id;
     const currentDeadline = task.deadline;
@@ -184,8 +189,13 @@ export async function buildTaskSummary(task) {
     buildPrioritySelectOptions(currentPriority, currentPriorityId);
 }
 
-// helper functions for task summary container
+/*
+The below are helper functions used to build the contents of the task summary panel.
+*/
 function buildTitleDiv(currentTask) {
+    /*
+    This function will build the div where a user can update a task's title in the task summary panel.
+    */
     const titleDiv = document.createElement('div');
     titleDiv.setAttribute('id', 'title-div');
     titleDiv.innerHTML = `
@@ -195,6 +205,9 @@ function buildTitleDiv(currentTask) {
 }
 
 export function buildDeadlineDiv(currentDeadline) {
+    /*
+    This function will build the div where a user can update a task's deadline in the task summary panel.
+    */
     const today = getDate();
     let deadline = '';
 
@@ -214,9 +227,12 @@ export function buildDeadlineDiv(currentDeadline) {
 }
 
 function buildListDiv(currentListId, currentList) {
+    /*
+    This function will build the div in the task summary panel where a user can update a task's list.
+    */
     const listDiv = document.createElement('div');
-    // listDiv.addEventListener('click', showCreateList);
     listDiv.setAttribute('id', 'list-div');
+
     listDiv.innerHTML = `
     <div id="summary-list">List</div>
     <select id="summary-list-select" class="summary-inp">
@@ -227,8 +243,12 @@ function buildListDiv(currentListId, currentList) {
 }
 
 function buildPriorityDiv(currentPriorityId, currentPriority) {
+    /*
+    This function will build the div in the task summary panel where a user can update a task's priority.
+    */
     const priorityDiv = document.createElement('div');
     priorityDiv.setAttribute('id', 'priority-div');
+
     priorityDiv.innerHTML = `
         <div id="summary-priority">Priority</div>
         <select id="summary-priority-select" class="summary-inp">
@@ -238,6 +258,10 @@ function buildPriorityDiv(currentPriorityId, currentPriority) {
 }
 
 function buildDescDiv(currentDesc) {
+    /*
+    This function will build the div in the task summary panel where a user can update their
+    task description via textarea.
+     */
     const descDiv = document.createElement('div');
     descDiv.setAttribute('id', 'desc-div');
     let descText = '';
@@ -253,6 +277,9 @@ function buildDescDiv(currentDesc) {
 };
 
 export async function buildListSelectOptions(currentListId, currentList) {
+    /*
+    This function will build the dropdown options in the task summary panel for updating a task's list.
+    */
     const listsRes = await fetch(`/api/lists`);
     const { lists } = await listsRes.json();
     const listOptions = document.querySelector('#summary-list-select');
@@ -263,12 +290,13 @@ export async function buildListSelectOptions(currentListId, currentList) {
     createListOpt.setAttribute('value', 'create-new');
     createListOpt.innerText = 'Create New';
 
-    // createListOpt.addEventListener('change', showCreateList);
-
     listOptions.appendChild(createListOpt);
 }
 
 export async function buildPrioritySelectOptions(currentPriority, currentPriorityId) {
+    /*
+    This function will build the dropdown options in the task summary panel for updating a task's priority.
+    */
     const priorityRes = await fetch(`/api/priorities`);
     const { priorities } = await priorityRes.json();
     const priorityOptions = document.querySelector('#summary-priority-select');
@@ -277,6 +305,9 @@ export async function buildPrioritySelectOptions(currentPriority, currentPriorit
 }
 
 function populateSelectOptions(table, currentSelectionName, selectHTMLElementName) {
+    /*
+    This function dropdown options and append them to the passed in html element.
+    */
     table.forEach(element => {
         if (element.name !== currentSelectionName) {
             const option = document.createElement('option');
@@ -287,8 +318,10 @@ function populateSelectOptions(table, currentSelectionName, selectHTMLElementNam
     });
 }
 
-// date functions
 export function getDate(day) {
+    /*
+    This function takes the date object and converts dates deadlines to use in the calendar input.
+    */
     let getDay;
     if (day) getDay = new Date(day);
     else getDay = new Date()
@@ -304,6 +337,9 @@ export function getDate(day) {
 }
 
 export function setTaskDeadline(taskDeadline) {
+    /*
+    This function will show the updated deadline for tasks in the inbox.
+    */
     const options = { weekday: 'long', year: 'numeric', month: 'short', day: 'numeric' };
     let deadline = getDate(taskDeadline);
     let today = getDate();
