+++ public/javascripts/dashboard-tasks.js
@@ -1,7 +1,7 @@
-import { updateTaskStatus } from './dashboard-recap.js'
-import { updatePriorityTag, updateDeadlineTag, changeTaskDeadline, moveTaskToNewList, moveTaskFromList, moveTaskFromTodayOrTomorrow } from './dashboard-summary.js'
+import { changeTaskDeadline, moveTaskToNewList } from './dashboard-summary.js'
 import { getDate, buildPrioritySelectOptions, decorateTaskWithDeadline } from './create-dom-elements.js';
 import { hideDivContainer, hideTaskSummary, selectNewList } from './display.js'
+import { updatePriorityTag, updateDeadlineTag, moveTaskFromList, moveTaskFromTodayOrTomorrow, updateTaskStatus } from './display-task-updates.js';
 
 export const checkAllBoxes = (e) => {
     const checkBox = document.querySelector('.checkbox-all > input');
