--- public/javascripts/stylesheets/main.css
@@ -1,233 +1,0 @@
-/* main container elements */
-#main #add-task-form {
-    padding: 10px 0 10px;
-    border-bottom: 1px dotted lightgray;
-}
-#main form {
-    display: flex;
-    flex-direction: column;
-}
-
-#main form .inp-div {
-    background-color: white;
-    z-index: 1;
-}
-
-#main form input {
-    width: 100%;
-}
-
-#add-task {
-    /* height: inherit; */
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    /* margin-bottom: 80px; */
-}
-
-.add-task-button {
-    align-self: flex-end;
-    margin-top: -40px;
-    transition: all 0.4s;
-}
-
-.add-task-button button {
-    color: #ffffff;
-    font-weight: bold;
-    background-color: var(--main-color);
-    border-radius: 3px;
-    border: none;
-}
-
-.add-task-button-transition {
-    margin-top: 0px;
-    transition: all 0.4s;
-}
-
-#add-task-input label {
-    z-index: 1;
-}
-
-#tasksContainer {
-    position: relative;
-    overflow-y: scroll;
-    padding-bottom: 80px;
-}
-
-::-webkit-scrollbar {
-    width: 0px;
-    background: transparent;
-}
-
-#tasksContainer div {
-    padding: 0 3px;
-    border-bottom: 1px dotted lightgray;
-}
-
-.single-task {
-    display: flex;
-    align-items: center;
-    height: 34px;
-    overflow: hidden;
-}
-
-.single-task label {
-    padding-left: 8px;
-}
-
-.single-task-selected {
-    background-color: #ffffc9;
-}
-
-.task {
-    display: flex;
-    align-items: center;
-}
-
-.single-task:hover {
-    background-color: #ffffc9;
-}
-
-.task > label {
-    padding: 0 8px;
-}
-
-/* add-task css needs lovin */
-#add-list,
-#rename-list {
-    display: flex;
-    border: 1px solid #d1d1d1;
-    border-radius: 5px;
-    top: 30%;
-    left: 30%;
-    box-shadow: 0 6px 12px rgb(0 0 0 / 10%);
-    background-color: white;
-    background-clip: padding-box;
-    padding: 29px 37px;
-    border-radius: 5px;
-    width: 390px;
-    height: 211px;
-    margin-top: 30px;
-    z-index: 10;
-}
-
-.add-task-box {
-    /* border: 1px red solid; */
-    margin-top: 5px;
-}
-
-#add-list-form {
-    display: flex;
-    flex-wrap: wrap;
-}
-
-#add-list-form label,
-#rename-list-form label {
-    padding-top: 12px;
-}
-
-#add-list-form input,
-#rename-list-form input {
-    align-self: center;
-    justify-self: center;
-    margin-top: 0;
-    /* top: 70px; */
-    right: 60px;
-    width: 315px;
-}
-
-#addList {
-    height: 28px;
-}
-
-.add-list-buttons {
-    display: flex;
-    flex-direction: row;
-    padding-top: 10px;
-    margin-left: -2px;
-}
-
-.add-list-buttons > button {
-    width: 70px;
-    height: 32px;
-    margin: 3px;
-}
-
-.cancel-submit-list {
-    background-color: #fbfbfa;
-    color: #444444;
-    /* box-shadow: 0 2px 6px rgb(0 0 0 / 10%); */
-    border: 1px solid #d1d1d1;
-}
-
-.add-task-box p {
-    /* border: red 1px solid; */
-    font-size: 16px;
-}
-
-.close {
-    /* background-image: url("../images/close-x.png");
-    background-repeat: no-repeat; */
-    position: absolute;
-    right: 15px;
-    top: 15px;
-    /* height: 20px;
-    width: 20px; */
-    color: rgb(211, 211, 211);
-    background-color: white;
-}
-
-.close:hover {
-    color: rgb(168, 168, 168);
-}
-
-/* date and priority tags for task items */
-.deadline-tag {
-    margin-left: auto;
-    margin-right: 5px;
-}
-
-.deadline-today {
-    color: var(--tasks-due-today);
-    font-weight: 600;
-}
-.deadline-overdue {
-    color: #bb0000;
-}
-
-.deadline-none {
-    display: none;
-}
-
-.deadline-soon {
-    color: gray;
-}
-
-.priority-tag {
-    margin: 7px 5px 6px;
-    padding: 3px;
-    color: white;
-    font-size: 9px;
-    text-transform: uppercase;
-    border-radius: 2px;
-}
-
-.priority-High {
-    background-color: #b548b5;
-}
-
-.priority-Medium {
-    background-color: rgb(228, 182, 57);
-}
-
-.priority-Low {
-    background-color: rgb(36, 206, 169);
-}
-
-.logout:hover {
-    background-color: #032c61;
-}
-
-.logout:active {
-    background-color: #03162f;
-}
