--- public/javascripts/stylesheets/sidebar.css
@@ -1,163 +1,0 @@
-/* sidebar elements */
-.sidebar-logo {
-    background-image: url("../images/logo-o-dark.png");
-    background-repeat: no-repeat;
-    background-size: 150px auto;
-    height: 100px;
-    width: 100%;
-}
-
-li {
-    list-style-type: none;
-}
-
-.lists {
-    margin-bottom: 10px;
-    padding: 0 0 10px 0;
-    border-bottom: 1px solid var(--divider-color);
-}
-
-/* here */
-.list-container {
-    display: none;
-}
-
-.lists {
-    line-height: 2em;
-    /* min-height: 500px; */
-}
-
-.list-item,
-.priority-item,
-.inbox-list {
-    padding-left: 17px;
-    color: rgb(66, 64, 64)
-}
-
-.selected-list {
-    color: #0060bf;
-    font-weight: bold;
-}
-
-.inbox-list:hover,
-.list-edit-option:hover,
-.list-header-container:hover,
-.sidebar-box:hover,
-.search-rec:hover {
-    background-color: var(--divider-color);
-}
-
-
-
-/* original lis turned into divs */
-ol li {
-    line-height: 2em;
-}
-
-.list-header-container {
-    display: grid;
-    grid-template-columns: 12px 1fr 12px;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.list-header {
-    /* padding-left: 4px; */
-    justify-self: start;
-    padding-left: 3px;
-}
-
-.sidebar-box {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    height: 22px;
-}
-
-.sidebar-icons {
-    display: none;
-    color: rgb(117, 165, 228);
-    padding-right: 1px;
-}
-
-.sidebar-box:hover > .sidebar-icons {
-    display: block;
-    /* color: red; */
-}
-
-.list-edit-dropdown {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    left: 40px;
-    width: 120px;
-    top: 15px;
-    background-color: white;
-    z-index: 3;
-    box-shadow: 0 1px 4px 1px rgb(0 0 0 / 20%);
-    border-radius: 5px;
-}
-
-.list-edit-option {
-    width: 100%;
-    padding-left: 8px;
-    /* border: 1px solid #a5c9e7; */
-    border-radius: 3px;
-}
-
-/* .add-list-button {
-    z-index: 10;
-    height: 16px;
-    width: 16px;
-    bottom: 5px;
-} */
-
-.fa-caret-square-down {
-    z-index: 10;
-}
-/*
-.fa-caret-right {
-    width: 17px;
-} */
-
-.fa-caret-right {
-    width: 17px;
-    transform: rotate(0deg);
-    transition: 0.2s linear;
-    transform-origin: 35% 40%;
-    padding-left: 5px;
-}
-
-.fa-caret-down {
-    transform: rotate(90deg);
-    transition: 0.2s linear;
-    /* margin-right: 10px; */
-    transform-origin: 35% 40%;
-    /* padding-left: 50px; */
-}
-
-.fa-plus-square {
-    color: #67a3d4;
-}
-
-.fa-plus-square:hover {
-    color: #478fca;
-    background-color: #a5c9e7;
-}
-
-.fa-caret-square-down:hover {
-    color: #478fca;
-    background-color: #a5c9e7;
-    z-index: 10;
-}
-
-.page-cover {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    /* z-index: -1; */
-    z-index: 6;
-    background-color: rgba(177, 171, 171, 0.5)
-}
